reorganize frontend type helpers and definitions
