+++ frontend/src/utils/astTypeDefinitions.ts
@@ -1,24 +1,4 @@
-// Comprehensive Luau AST Type Definitions
-// Generated from official Luau type definitions for maximum accuracy
-
-export interface PropertyDefinition {
-  name: string;
-  type: string | string[]; // string[] for unions like "true" | "false"
-  optional?: boolean;
-  generic?: string; // for Token<"specific"> types
-}
-
-export interface ASTTypeDefinition {
-  properties?: PropertyDefinition[];
-  kinds?: Record<string, ASTTypeDefinition>;
-  baseType?: string; // for intersections like "Token &"
-  unionMembers?: string[]; // for "AstExpr = A | B | C"
-}
-
-export interface GenericTypeDefinition {
-  baseType: string;
-  genericType: string;
-}
+import { ASTTypeDefinition } from "./astTypeHelpers";
 
 export const astTypeDefinitions: Record<string, ASTTypeDefinition> = {
   // === UTILITY TYPES ===
@@ -1020,162 +1000,4 @@ export const astTypeDefinitions: Record<string, ASTTypeDefinition> = {
       "AstTypePackVariadic",
     ],
   },
-};
-
-export function getTypeDefinition(
-  typeName: string | GenericTypeDefinition
-): ASTTypeDefinition | undefined {
-  if (typeof typeName === "string") {
-    return astTypeDefinitions[typeName];
-  }
-
-  return getGenericTypeDefinition(typeName);
-}
-
-export function getGenericTypeDefinition(
-  genericType: GenericTypeDefinition
-): ASTTypeDefinition {
-  switch (genericType.baseType) {
-    case "Pair":
-      const splitGenericType = genericType.genericType.split(",");
-      return {
-        properties: [
-          { name: "node", type: splitGenericType[0] },
-          {
-            name: "separator",
-            type: `Token${
-              splitGenericType[1] ? `<${splitGenericType[1].trim()}>` : ""
-            }`,
-            optional: true,
-          },
-        ],
-      };
-    case "Punctuated":
-      return {
-        properties: [
-          { name: "", type: `{ Pair<${genericType.genericType}> }` },
-        ],
-      };
-    default:
-      return { properties: [] };
-  }
-}
-
-export function getAllTypes(): string[] {
-  return Object.keys(astTypeDefinitions);
-}
-
-export function getAllNodeKeys(): string[] {
-  // Extract all unique property names from all type definitions
-  const allProperties: string[] = [];
-
-  Object.values(astTypeDefinitions).forEach((typeDef) => {
-    // Add properties from the main properties array
-    if (typeDef.properties) {
-      typeDef.properties.forEach((prop) => {
-        allProperties.push(prop.name);
-      });
-    }
-
-    // Add properties from kinds (for union types)
-    if (typeDef.kinds) {
-      Object.values(typeDef.kinds).forEach((kindDef) => {
-        if (kindDef.properties) {
-          kindDef.properties.forEach((prop) => {
-            allProperties.push(prop.name);
-          });
-        }
-      });
-    }
-  });
-
-  return Array.from(new Set(allProperties)).sort();
-}
-
-const resolveEntriesType = (value: any[]): string => {
-  if (value[0] && value[0].colon) {
-    return "{ AstTypeTableItem }";
-  }
-
-  return "{ AstExprTableItem }";
-};
-
-const resolveEntriesKind = (value: any[]): string => {
-  return value.length > 0 ? value[0].kind : "";
-};
-
-const arrayTypeFallbacks: Record<string, string | ((item: any[]) => string)> = {
-  statements: "{ AstStat }",
-  leadingTrivia: "{ Trivia }",
-  trailingTrivia: "{ Trivia }",
-  attributes: "{ AstAttribute }",
-  expressions: "{ AstExpr }",
-  elseifs: "{ AstExprIfElseIfs }",
-  entries: resolveEntriesType,
-};
-
-export const getArrayType = (
-  arrayKey: string,
-  arrayValue?: any[]
-): [string, string] => {
-  const fallback = arrayTypeFallbacks[arrayKey];
-  if (typeof fallback === "function") {
-    const entryKind = resolveEntriesKind(arrayValue ?? []);
-    return [fallback(arrayValue ?? []), entryKind];
-  }
-  return [fallback, ""];
-};
-
-// true if type is a table of subtypes (i.e. { Trivia })
-export const isArrayType = (type: string) => {
-  return /^\{\s*[\w<>| ]+\s*\}$/.test(type);
-};
-
-export const unpackArrayType = (type: string): string => {
-  // Match any content inside curly braces, including nested angle brackets, quotes, etc.
-  // This regex matches everything between the first { and last }
-  const matchAny = type.match(/^\{\s*([\s\S]+?)\s*\}$/);
-  return matchAny ? matchAny[1] : type;
-};
-
-// Parse generic types like "Pair<AstExpr>" -> { baseType: "Pair", genericType: "AstExpr" }
-export const parseGenericType = (type: string) => {
-  const match = type.match(/^(\w+)<(.+)>$/);
-  if (match) {
-    return { baseType: match[1], genericType: match[2] };
-  }
-  return null;
-};
-
-export const getType = (
-  value: any,
-  nodeKey: string,
-  parentInferredType?: string | string[]
-) => {
-  let type = value._astType;
-  let kind = value.kind;
-  if (Array.isArray(value)) {
-    [type, kind] = getArrayType(
-      nodeKey,
-      nodeKey === "entries" ? value : undefined
-    );
-  }
-  if (parentInferredType != type) {
-    console.log("type:", type, "vs. parentInferredType:", parentInferredType);
-  }
-  type = !type ? parentInferredType : type; // if type is null, fallback to parentInferredType (this should handle Punctuated well)
-  return [type, kind];
-};
-
-export const getTypeMetadata = (type: string): [ASTTypeDefinition | undefined, boolean] => {
-  const genericTypeInfo = Array.isArray(type)
-    ? undefined
-    : parseGenericType(type);
-  const isArray = isArrayType(type);
-  const unpackedType = isArray
-    ? unpackArrayType(type)
-    : genericTypeInfo
-    ? genericTypeInfo
-    : type;
-  return [getTypeDefinition(unpackedType), isArray];
 };
